+++ NewFront-end/Domicilios/src/components/layout/Header.jsx
@@ -4,6 +4,7 @@ import LogoutButton from '../navegacion/LogoutButton';
 import { FaBell, FaBars, FaTimes } from "react-icons/fa";
 import NotificacionesBell from '../notificaciones/NotificacionesBell';
 import { useSolicitudes } from '../../services/SolicitudesProvider';
+import axios from 'axios';
 
 const Header = () => {
   const [showNotifications, setShowNotifications] = useState(false);
@@ -11,10 +12,15 @@ const Header = () => {
   const userType = localStorage.getItem('userType');
   const { listarSolicitudes, solicitudes = [] } = useSolicitudes() || {};
   const [notificacionesPendientes, setNotificacionesPendientes] = useState(0);
+  const [disponibilidadActual, setDisponibilidadActual] = useState('disponible');
+  const [disponibilidad, setDisponibilidad] = useState('disponible');
 
   useEffect(() => {
     if (listarSolicitudes) {
       listarSolicitudes();
+    }
+    if (userType === 'domiciliario') {
+      getDisponibilidad();
     }
     const interval = setInterval(() => {
       if (listarSolicitudes) {
@@ -29,6 +35,34 @@ const Header = () => {
     const pendientes = solicitudes.filter(solicitud => solicitud.estado === 'en_curso');
     setNotificacionesPendientes(pendientes.length);
   }, [solicitudes]);
+
+  const getDisponibilidad = async () => {
+    const id_usuario = localStorage.getItem('userId');
+    const respuesta = await axios.get(`${import.meta.env.VITE_API_URL}domiciliario/consultar/${id_usuario}`);
+    setDisponibilidad(respuesta.data[0].disponibilidad);
+    setDisponibilidadActual(respuesta.data[0].disponibilidad);
+  };
+
+  const handleDisponibilidadChange = async () => {
+    try {
+      const id_usuario = localStorage.getItem('userId');
+      const response = await fetch(`${import.meta.env.VITE_API_URL}domiciliario/disponibilidad/${id_usuario}`, {
+        method: 'PUT',
+        headers: { 'Content-Type': 'application/json' }
+      });
+
+      if (response.ok) {
+        const data = await response.json();
+        console.log('Nueva disponibilidad:', data.nuevaDisponibilidad);
+        setDisponibilidadActual(data.nuevaDisponibilidad);
+        if (listarSolicitudes) {
+          await listarSolicitudes();
+        }
+      }
+    } catch (error) {
+      console.error('Error al actualizar disponibilidad:', error);
+    }
+  };
 
   const toggleNotifications = () => {
     setShowNotifications(!showNotifications);
@@ -48,8 +82,7 @@ const Header = () => {
         </div>
 
         <div className="flex gap-6 items-center">
-          {/* Menú de navegación */}
-          <ul className="hidden sm:flex  space-x-4">
+          <ul className="hidden sm:flex space-x-4">
             <Link to="/Home" className="text-gray-800 hover:text-green-600">Home</Link>
             {userType === 'domiciliario' && (
               <>
@@ -61,7 +94,8 @@ const Header = () => {
               <Link to="/PanelDeControl" className="text-gray-800 hover:text-green-600">Panel De Control</Link>
             )}
           </ul>
-          {(userType === 'domiciliario' ||userType === 'administrador' )&& (
+          
+          {(userType === 'domiciliario' || userType === 'administrador') && (
             <div className="sm:hidden">
               <button onClick={toggleMenu} className="text-3xl">
                 {isMenuOpen ? <FaTimes /> : <FaBars />}
@@ -72,14 +106,35 @@ const Header = () => {
                   <Link to="/novedades" className="block text-gray-800 hover:text-green-600 mb-2">Registrar novedad</Link>
                   <Link to="/NotificacionesDom" className="block text-gray-800 hover:text-green-600 mb-2">Domicilios</Link>
                   {userType === 'administrador' && (
-                  <Link to="/PanelDeControl" className="text-gray-800 hover:text-green-600">Panel De Control</Link>
-                )}
+                    <Link to="/PanelDeControl" className="text-gray-800 hover:text-green-600">Panel De Control</Link>
+                  )}
                 </div>
               )}
             </div>
           )}
+
           {userType === 'domiciliario' && (
             <>
+              <button
+                onClick={handleDisponibilidadChange}
+                className={`px-4 py-2 rounded-full transition-colors duration-50 
+                  ${disponibilidadActual === 'disponible'
+                    ? 'bg-blue-300 text-blue-900 hover:bg-blue-400'
+                    : disponibilidadActual === 'no_disponible'
+                      ? 'bg-red-300 text-red-900 hover:bg-red-400'
+                      : 'bg-gray-300 text-gray-900 hover:bg-gray-400'
+                } border border-black`}
+              >
+                <div className="flex items-center space-x-2">
+                  <div className={`w-2 h-2 rounded-full ${
+                    disponibilidadActual === 'disponible' ? 'bg-blue-600'
+                    : disponibilidadActual === 'no_disponible' ? 'bg-red-600'
+                    : 'bg-gray-600'
+                  }`}></div>
+                  <span>{disponibilidadActual}</span>
+                </div>
+              </button>
+
               <button className="relative" onClick={toggleNotifications}>
                 <FaBell className="text-3xl cursor-pointer" />
                 {notificacionesPendientes > 0 && (
